+++ src/views/index.ts
@@ -4,3 +4,4 @@
 
 export { SquadTreeProvider, SquadTreeItem } from './SquadTreeProvider';
 export { WorkDetailsWebview } from './WorkDetailsWebview';
+export { IssueDetailWebview } from './IssueDetailWebview';